[fix]clear registry

+++ rpc-framework-simple/src/main/java/github/javaguide/registry/zk/util/CuratorUtils.java
@@ -2,6 +2,7 @@ package github.javaguide.registry.zk.util;
 
 import github.javaguide.enums.RpcConfigEnum;
 import github.javaguide.utils.PropertiesFileUtil;
+import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.curator.RetryPolicy;
 import org.apache.curator.framework.CuratorFramework;
@@ -12,6 +13,8 @@ import org.apache.curator.framework.recipes.cache.PathChildrenCacheListener;
 import org.apache.curator.retry.ExponentialBackoffRetry;
 import org.apache.zookeeper.CreateMode;
 
+import java.net.InetAddress;
+import java.net.InetSocketAddress;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
@@ -83,10 +86,12 @@ public final class CuratorUtils {
     /**
      * Empty the registry of data
      */
-    public static void clearRegistry(CuratorFramework zkClient) {
+    public static void clearRegistry(CuratorFramework zkClient, InetSocketAddress inetSocketAddress) {
         REGISTERED_PATH_SET.stream().parallel().forEach(p -> {
             try {
-                zkClient.delete().forPath(p);
+                if (p.endsWith(inetSocketAddress.toString())) {
+                    zkClient.delete().forPath(p);
+                }
             } catch (Exception e) {
                 log.error("clear registry for path [{}] fail", p);
             }
